1.Removed styling for .App; 2.Removed blank pages; 3.Keep all file extensions to be js rather than jsx.
