+++ backend/server.js
@@ -16,7 +16,7 @@ const port = process.env.PORT || 3000;
 app.use(cors());
 app.use(morgan("combined"));
 app.use(express.json());
-app.use(express.urlencoded({ extended: false }));
+app.use(express.urlencoded({ extended: true }));
 
 app.use("/auth", userAuthRoutes);
 app.use("/api/users", userRoutes);
