+++ frontend/src/App.css
@@ -1,10 +1,5 @@
 .App {
   text-align: center;
-  min-height: 100vh;
-  align-items: center;
-  justify-content: center;
-  background-image: linear-gradient(370deg, orange 50%, cyan);
-  color: white;
 }
 
 .App-logo {
@@ -52,7 +47,7 @@
   margin: 2rem;
   padding: 5rem;
   border: 1px;
-  border-radius: 10px solid white;
+  border-radius: 10px solid black;
 }
 
 label {
@@ -63,13 +58,12 @@ label {
 input {
   margin: 0.5rem 0;
   padding: 1rem;
-  border: none;
   border-radius: 10px;
 }
 
 button {
   border: none;
-  background-color: white;
+  background-color: gray;
   padding: 20px;
   border-radius: 10px;
   cursor: pointer;
