--- frontend/src/pages/PasswordPage1.js
@@ -1,10 +1,0 @@
-import * as React from 'react'
-import { Password1 } from '../jsx/Password1'
-
-function App() {
-    return (
-    <Password1 />
-    )
-}
-
-export default App
