--- frontend/src/pages/PasswordPage2.js
@@ -1,10 +1,0 @@
-import * as React from 'react'
-import { Password2 } from '../jsx/Password2'
-
-function App() {
-    return (
-    <Password2 />
-    )
-}
-
-export default App
