+++ frontend/src/pages/PasswordResetPage.js
@@ -1,9 +1,7 @@
 import * as React from 'react'
 import {useState} from 'react'
-import Button from '@mui/material/Button'
 
-function Login() {
-    const [email, setEmail] = useState('')
+function App() {
     const [pass, setPass] = useState('')
 
     const handleSubmit = (e) => {
@@ -16,18 +14,18 @@ function Login() {
           <h1 className="card-title">Log in</h1>
           <form className="login-form" onSubmit={handleSubmit}>
             <div className="form-group">
-              <label htmlFor="InputEmail">Email</label>
+              <label htmlFor="InputPassword">New password</label>
               <input
-                type="email"
+                type="password"
                 className="form-control"
-                id="InputEmail"
-                placeholder='email@gmail.com'
-                value={email}
-                onChange={(e) => setEmail(e.target.value)}
+                id="InputPassword"
+                placeholder='***********'
+                value={pass}
+                onChange={(e) => setPass(e.target.value)}
               />
             </div>
             <div className="form-group">
-              <label htmlFor="InputPassword">Password</label>
+              <label htmlFor="InputPassword">Comfirm password</label>
               <input
                 type="password"
                 className="form-control"
@@ -41,14 +39,12 @@ function Login() {
               type="submit"
               className="btn"
             >
-              LOG IN
+              CONFIRM
             </button>
           </form>
-          <Button href='passwordEmailLink'>Forgot password?</Button>
-          <Button href='signup'>Don't have an account? Sign up HERE</Button>
         </div>
     </div>
     )
 }
 
-export default Login
+export default App
